+++ javacv/src/main/java/org/javacv/face/recognition/FaceRecognition.java
@@ -1,4 +1,4 @@
-package org.javacv.face.image;
+package org.javacv.face.recognition;
 
 import java.io.File;
 import org.bytedeco.javacpp.opencv_face.FaceRecognizer;
