+++ javacv/src/main/java/org/javacv/face/recognition/GenderTrainer.java
@@ -1,4 +1,4 @@
-package org.javacv.face.image;
+package org.javacv.face.recognition;
 
 import com.google.gson.Gson;
 import java.io.File;
