+++ javacv/src/main/java/org/javacv/face/recognition/RecognizerType.java
@@ -1,4 +1,4 @@
-package org.javacv.face.image;
+package org.javacv.face.recognition;
 
 /**
  *
