+++ javacv/src/main/java/org/javacv/fusion/AffineAlignmentProcess.java
@@ -1,4 +1,4 @@
-package org.javacv.img;
+package org.javacv.fusion;
 
 import java.util.ArrayList;
 import java.util.List;
