+++ javacv/src/main/java/org/javacv/ui/CanvasDemo.java
@@ -7,10 +7,10 @@ import java.util.concurrent.Executors;
 import javax.swing.JFrame;
 
 import org.bytedeco.javacv.CanvasFrame;
-import org.javacv.face.image.FaceRecognition;
-import org.javacv.face.image.DetectorService;
-import org.javacv.face.image.GenderTrainer;
-import org.javacv.face.image.RecognizerType;
+import org.javacv.face.recognition.FaceRecognition;
+import org.javacv.face.detection.DetectorService;
+import org.javacv.face.recognition.GenderTrainer;
+import org.javacv.face.recognition.RecognizerType;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
