+++ javacv/src/main/resources/org/javacv/face/detection/haarcascade_frontalface_alt_tree.xml
