+++ javacv/src/test/java/org/javacv/fusion/AffineAlignmentProcessTest.java
@@ -1,4 +1,4 @@
-package org.javacv.img;
+package org.javacv.fusion;
 
 import java.util.Arrays;
 import java.util.List;
