+++ javacv/src/test/resources/org/javacv/face/detection/face.jpg
