+++ javacv/src/test/resources/org/javacv/face/detection/squad.jpg
