+++ javacv/src/test/resources/org/javacv/face/detection/tree.jpg
