+++ javacv/src/test/resources/org/javacv/face/recognition/1_test.jpg
