+++ javacv/src/test/resources/org/javacv/face/recognition/2_test.jpg
