+++ javacv/src/test/resources/org/javacv/face/recognition/3_salma_test.jpg
