+++ javacv/src/test/resources/org/javacv/face/recognition/back.jpg
