+++ javacv/src/test/resources/org/javacv/fusion/Picture_201508010708_0.jpg
