+++ javacv/src/test/resources/org/javacv/fusion/memorial0061.png
