+++ javacv/src/test/resources/org/javacv/fusion/memorial0064.png
