+++ javacv/src/test/resources/org/javacv/fusion/memorial0067.png
