+++ javacv/src/test/resources/org/javacv/fusion/stat_1.jpg
