+++ javacv/src/test/resources/org/javacv/fusion/stat_2.jpg
