+++ javacv/src/test/resources/org/javacv/fusion/stat_3.jpg
